+++ backend/src/repository/stats/__init__.py
@@ -1,0 +1,29 @@
+from repository.stats.orchestration import (
+    compute_check_ratio,
+    compute_rec_fail_ratio,
+    get_bot_general_stats,
+    get_bot_user_stats,
+    get_cumulative_records,
+    get_cumulative_volunteers,
+    get_progress,
+    get_project_statistics,
+    get_user_by_id,
+    get_user_by_name,
+    get_user_statistics,
+    get_volunteers_achievements,
+)
+
+__all__ = [
+    "compute_check_ratio",
+    "compute_rec_fail_ratio",
+    "get_bot_general_stats",
+    "get_bot_user_stats",
+    "get_cumulative_records",
+    "get_cumulative_volunteers",
+    "get_progress",
+    "get_project_statistics",
+    "get_user_by_id",
+    "get_user_by_name",
+    "get_user_statistics",
+    "get_volunteers_achievements",
+]
